integration-tests: rename relocations to btf_relocations
In preparation of adding actual ELF relocation tests
