+++ test/integration-test/src/tests/mod.rs
@@ -4,10 +4,10 @@ use libc::{uname, utsname};
 use regex::Regex;
 use std::{ffi::CStr, mem};
 
+pub mod btf_relocations;
 pub mod elf;
 pub mod load;
 pub mod rbpf;
-pub mod relocations;
 pub mod smoke;
 
 pub use integration_test_macros::integration_test;
